application/requirements.txt
Flask
python-dotenv
pytz
requests
pipreqs